factor out nft sizing logic
