+++ apps/mobile/src/screens/NftDetailScreen/NftDetailAsset/NftDetailAsset.tsx
@@ -1,20 +1,19 @@
-import { useCallback, useMemo, useState } from 'react';
-import { LayoutChangeEvent, useWindowDimensions, View, ViewProps } from 'react-native';
+import { useCallback, useContext, useMemo } from 'react';
+import { View, ViewProps } from 'react-native';
 import SkeletonPlaceholder from 'react-native-skeleton-placeholder';
 import { useFragment } from 'react-relay';
 import { graphql } from 'relay-runtime';
 
 import { GallerySkeleton } from '~/components/GallerySkeleton';
 import { NftDetailAssetFragment$key } from '~/generated/NftDetailAssetFragment.graphql';
-import { fitDimensionsToContainerContain } from '~/screens/NftDetailScreen/NftDetailAsset/fitDimensionToContainer';
+import { NftDetailAssetCacheSwapperContext } from '~/screens/NftDetailScreen/NftDetailAsset/NftDetailAssetCacheSwapper';
 import { NftDetailAssetHtml } from '~/screens/NftDetailScreen/NftDetailAsset/NftDetailAssetHtml';
 import { NftDetailAssetImage } from '~/screens/NftDetailScreen/NftDetailAsset/NftDetailAssetImage';
 import { NftDetailAssetVideo } from '~/screens/NftDetailScreen/NftDetailAsset/NftDetailAssetVideo';
 import { Dimensions } from '~/screens/NftDetailScreen/NftDetailAsset/types';
+import { useNftDetailAssetSizer } from '~/screens/NftDetailScreen/NftDetailAsset/useNftDetailAssetSizer';
 import { CouldNotRenderNftError } from '~/shared/errors/CouldNotRenderNftError';
 import getVideoOrImageUrlForNftPreview from '~/shared/relay/getVideoOrImageUrlForNftPreview';
-
-type ImageState = { kind: 'loading' } | { kind: 'loaded'; dimensions: Dimensions | null };
 
 type NftDetailProps = {
   tokenRef: NftDetailAssetFragment$key;
@@ -61,44 +60,17 @@ export function NftDetailAsset({ tokenRef, style }: NftDetailProps) {
     tokenRef
   );
 
-  const windowDimensions = useWindowDimensions();
+  const assetSizer = useNftDetailAssetSizer();
 
-  const [imageState, setImageState] = useState<ImageState>({ kind: 'loading' });
-  const [viewDimensions, setViewDimensions] = useState<Dimensions | null>(null);
+  const cachedAssetSwapperContext = useContext(NftDetailAssetCacheSwapperContext);
+  const handleLoad = useCallback(
+    (dimensions?: Dimensions | null) => {
+      cachedAssetSwapperContext?.markDetailAssetAsLoaded();
 
-  const handleViewLayout = useCallback((event: LayoutChangeEvent) => {
-    const { width, height } = event.nativeEvent.layout;
-
-    setViewDimensions({ width, height });
-  }, []);
-
-  const handleLoad = useCallback((dimensions?: Dimensions | null) => {
-    setImageState({ kind: 'loaded', dimensions: dimensions ?? null });
-  }, []);
-
-  const finalAssetDimensions = useMemo((): Dimensions => {
-    if (viewDimensions && imageState.kind === 'loaded' && imageState.dimensions) {
-      // Give the piece a little bit of breathing room. This might be an issue
-      // if we ever support landscape view (turning your phone horizontally).
-      const MAX_HEIGHT = windowDimensions.height - 400;
-
-      // Width is the width of the parent view (the screen - some padding)
-      // Height is the max height for the image
-      //
-      // This will fit the image to the screen appropriately.
-      const containerDimensions: Dimensions = { width: viewDimensions.width, height: MAX_HEIGHT };
-
-      return fitDimensionsToContainerContain({
-        container: containerDimensions,
-        source: imageState.dimensions,
-      });
-    }
-
-    // This is a fallback for when we don't have the image dimensions yet.
-    // The user will never see the image in this state since it will be covered
-    // by a loading skeleton UI anyway.
-    return { width: 300, height: 300 };
-  }, [imageState, viewDimensions, windowDimensions.height]);
+      assetSizer.handleLoad(dimensions);
+    },
+    [assetSizer, cachedAssetSwapperContext]
+  );
 
   const inner = useMemo(() => {
     if (
@@ -120,7 +92,7 @@ export function NftDetailAsset({ tokenRef, style }: NftDetailProps) {
         <NftDetailAssetImage
           onLoad={handleLoad}
           imageUrl={imageUrl}
-          outputDimensions={finalAssetDimensions}
+          outputDimensions={assetSizer.finalAssetDimensions}
         />
       );
     } else if (token.media?.__typename === 'VideoMedia') {
@@ -134,7 +106,7 @@ export function NftDetailAsset({ tokenRef, style }: NftDetailProps) {
         <NftDetailAssetVideo
           onLoad={handleLoad}
           videoUrl={videoUrl}
-          outputDimensions={finalAssetDimensions}
+          outputDimensions={assetSizer.finalAssetDimensions}
         />
       );
     } else if (token.media?.__typename === 'HtmlMedia') {
@@ -153,23 +125,23 @@ export function NftDetailAsset({ tokenRef, style }: NftDetailProps) {
     throw new CouldNotRenderNftError('NftDetailAsset', 'Unsupported media type', {
       typename: token.media?.__typename,
     });
-  }, [finalAssetDimensions, handleLoad, token]);
+  }, [assetSizer.finalAssetDimensions, handleLoad, token]);
 
   return (
     <View
       style={style}
       className="relative flex flex-row justify-center"
-      onLayout={handleViewLayout}
+      onLayout={assetSizer.handleViewLayout}
     >
       {inner}
 
       {/* Show a skeleton placeholder over the image while it's loading */}
-      {imageState.kind === 'loading' && (
+      {assetSizer.imageState.kind === 'loading' && (
         <View className="absolute">
           <GallerySkeleton>
             <SkeletonPlaceholder.Item
-              width={viewDimensions?.width}
-              height={viewDimensions?.height}
+              width={assetSizer.viewDimensions?.width}
+              height={assetSizer.viewDimensions?.height}
             />
           </GallerySkeleton>
         </View>
